+++ src/io.c
@@ -1,0 +1,26 @@
+/* vi:set ts=8 sts=4 sw=4:
+ *
+ * VIM - Vi IMproved	by Bram Moolenaar
+ *
+ * Do ":help uganda"  in Vim to read copying and usage conditions.
+ * Do ":help credits" in Vim to see a list of people who contributed.
+ * See README.txt for an overview of the Vim source code.
+ */
+
+/*
+ * io.c -- filesystem access, event loop etc.
+ */
+
+#include "vim.h"
+
+#include "uv.h"
+
+int mch_chdir(char *path)
+{
+  if (p_verbose >= 5) {
+    verbose_enter();
+    smsg((char_u *)"chdir(%s)", path);
+    verbose_leave();
+  }
+  return uv_chdir(path);
+}
